refactor : pagenation 및 초대하기 기능 연동의 용이성을 위해 TableLayer를 각 Table 컴포넌트 내부에 적용
